Updates to the docking submodule.
