+++ covid_moonshot_ml/__init__.py
@@ -2,8 +2,10 @@
 
 # Add imports here
 # from .covid_moonshot_ml import *
-from . import nn
 from . import data
+from . import docking
+from . import nn
+from . import schema
 
 # Handle versioneer
 from ._version import get_versions
