+++ covid_moonshot_ml/docking/mcs.py
@@ -1,20 +1,7 @@
-import argparse
-import json
-import multiprocessing as mp
 import numpy as np
-from openeye import oechem, oedepict
-import pickle as pkl
-from rdkit import Chem
-from rdkit.Chem import rdFMCS, Draw
-from rdkit.Chem.Draw import rdMolDraw2D
-import time
-
-from run_docking import parse_xtal
-from schema import ExperimentalCompoundDataUpdate, CrystalCompoundData, \
-    EnantiomerPairList
 
 def rank_structures_openeye(exp_smi, exp_id, search_smis, search_ids,
-    smi_conv, str_based=False, out_fn=None, n_draw=0):
+    smi_conv=None, str_based=False, out_fn=None, n_draw=0):
     """
     Rank all molecules in search_mols based on their MCS with exp_mol.
 
@@ -26,6 +13,10 @@ def rank_structures_openeye(exp_smi, exp_id, search_smis, search_ids,
         Molecules generated from the SMILES of the ligands in the crystal
         compounds
     """
+    from openeye import oechem, oedepict
+
+    if smi_conv is None:
+        smi_conv = _smi_conv_oe
 
     if str_based:
         """
@@ -125,18 +116,10 @@ def rank_structures_openeye(exp_smi, exp_id, search_smis, search_ids,
             oedepict.OEWriteImage(f'{out_fn}_{search_ids[mol_idx]}_{i}.png',
                 image)
 
-    """
-    https://docs.eyesopen.com/toolkits/python/depicttk/molalign.html
-    https://docs.eyesopen.com/toolkits/python/oechemtk/patternmatch.html#section-patternmatch-mcss
-    https://docs.eyesopen.com/toolkits/python/oechemtk/OEChemConstants/OEExprOpts.html#OEChem::OEExprOpts::HvyDegree
-    https://docs.eyesopen.com/toolkits/python/oechemtk/OEChemClasses/OEMCSSearch.html
-    """
-
     return(sort_idx)
 
-
 def rank_structures_rdkit(exp_smi, exp_id, search_smis, search_ids,
-    smi_conv, str_based=False, out_fn=None, n_draw=0):
+    smi_conv=None, str_based=False, out_fn=None, n_draw=0):
     """
     Rank all molecules in search_mols based on their MCS with exp_mol.
 
@@ -148,14 +131,17 @@ def rank_structures_rdkit(exp_smi, exp_id, search_smis, search_ids,
         Molecules generated from the SMILES of the ligands in the crystal
         compounds
     """
+    from rdkit import Chem
+    from rdkit.Chem import rdFMCS, Draw
+    from rdkit.Chem.Draw import rdMolDraw2D
+
+    if smi_conv is None:
+        smi_conv = _smi_conv_rdkit
 
     if str_based:
         atom_compare = rdFMCS.AtomCompare.CompareAny
     else:
         atom_compare = rdFMCS.AtomCompare.CompareElements
-
-    print(f'Starting {exp_id}', flush=True)
-    start_time = time.time()
 
     exp_mol = smi_conv(exp_smi)
 
@@ -230,117 +216,16 @@ def rank_structures_rdkit(exp_smi, exp_id, search_smis, search_ids,
 
     Draw.MolToFile(exp_mol, f'{out_fn}.png')
 
-    end_time = time.time()
-    print(f'Finished {exp_id} ({end_time-start_time} s)', flush=True)
     return(sort_idx)
 
-def smi_conv_rdkit(s):
+def _smi_conv_rdkit(s):
+    from rdkit import Chem
+
     return(Chem.MolFromSmiles(Chem.CanonSmiles(s)))
 
-def smi_conv_oe(s):
+def _smi_conv_oe(s):
+    from openeye import oechem
+
     mol = oechem.OEGraphMol()
     oechem.OESmilesToMol(mol, s)
     return(mol)
-
-################################################################################
-def get_args():
-    parser = argparse.ArgumentParser(description='')
-
-    ## Input arguments
-    parser.add_argument('-exp', required=True,
-        help='JSON file giving experimental results.')
-    parser.add_argument('-x', required=True,
-        help='CSV file with crystal compound information.')
-    parser.add_argument('-x_dir', required=True,
-        help='Directory with crystal structures.')
-
-    ## Output arguments
-    parser.add_argument('-o', required=True, help='Main output directory.')
-
-    ## Performance arguments
-    parser.add_argument('-n', default=1, type=int,
-        help='Number of processors to use.')
-    parser.add_argument('-sys', default='rdkit',
-        help='Which package to use for MCS search [rdkit, oe].')
-    parser.add_argument('-str', action='store_true', help=('Use '
-        'structure-based matching instead of element-based matching for MCS.'))
-    parser.add_argument('-ep', action='store_true',
-        help='Input data is in EnantiomerPairList format.')
-
-    return(parser.parse_args())
-
-def main():
-    args = get_args()
-
-    ## Load all compounds with experimental data and filter to only achiral
-    ##  molecules (to start)
-    if args.ep:
-        exp_compounds = [c for ep in EnantiomerPairList(
-            **json.load(open(args.exp, 'r'))).pairs \
-            for c in (ep.active, ep.inactive)]
-    else:
-        exp_compounds = [c for c in ExperimentalCompoundDataUpdate(
-            **json.load(open(args.exp, 'r'))).compounds if c.smiles is not None]
-        exp_compounds = np.asarray([c for c in exp_compounds if c.achiral])
-
-    ## Find relevant crystal structures
-    xtal_compounds = parse_xtal(args.x, args.x_dir)
-
-    ## See if I can add a title to the MCS plots for the xtal id
-    compound_ids = [c.compound_id for c in exp_compounds]
-    xtal_ids = [x.dataset for x in xtal_compounds]
-    xtal_smiles = [x.smiles for x in xtal_compounds]
-
-    if args.sys.lower() == 'rdkit':
-        ## Convert SMILES to RDKit mol objects for MCS
-        ## Need to canonicalize SMILES first because RDKit MCS seems to have
-        ##  trouble with non-canon SMILES
-        smi_conv = smi_conv_rdkit
-        rank_fn = rank_structures_rdkit
-    elif args.sys.lower() == 'oe':
-        smi_conv = smi_conv_oe
-        rank_fn = rank_structures_openeye
-
-    # exp_lig_mols = [smi_conv(c.smiles) for c in exp_compounds]
-    # xtal_lig_mols = [smi_conv(x.smiles) for x in xtal_compounds]
-    # exp_lig_mols = np.asarray([
-    #     Chem.MolFromSmiles(Chem.CanonSmiles(c.smiles)) \
-    #     for c in exp_compounds])
-    # xtal_lig_mols = np.asarray([
-    #     Chem.MolFromSmiles(Chem.CanonSmiles(x.smiles)) \
-    #     for x in xtal_compounds])
-
-    print(f'{len(exp_compounds)} experimental compounds')
-    print(f'{len(xtal_compounds)} crystal structures')
-    print('Finding best docking structures', flush=True)
-    ## Prepare the arguments to pass to starmap
-    mp_args = [(c.smiles, c.compound_id, xtal_smiles, xtal_ids, smi_conv,
-        args.str, f'{args.o}/{c.compound_id}', 10) for c in exp_compounds]
-    # mp_args = [(m, c, xtal_lig_mols, xtal_ids, f'{args.o}/{c}', 10) \
-    #     for m, c in zip(exp_lig_mols, compound_ids)]
-    n_procs = min(args.n, mp.cpu_count(), len(exp_compounds))
-    with mp.Pool(n_procs) as pool:
-        res = pool.starmap(rank_fn, mp_args)
-    # res = [rank_fn(*a) for a in mp_args]
-
-    # ## Problem molecules
-    # c = 'RED-RED-10c9212c-16'
-    # m = [e for e in exp_compounds if e.compound_id == c][0]
-    # m = Chem.MolFromSmiles(Chem.CanonSmiles(m.smiles))
-
-    # xtal_id = 'Mpro-P1701'
-    # xtal_lig_mols = [x for x in xtal_compounds if xtal_id in x.dataset]
-    # xtal_lig_mols = [Chem.MolFromSmiles(Chem.CanonSmiles(x.smiles)) \
-    #     for x in xtal_lig_mols]
-
-    # print(Chem.CanonSmiles([e for e in exp_compounds if e.compound_id == c][0].smiles))
-    # print(Chem.CanonSmiles([x for x in xtal_compounds if xtal_id in x.dataset][0].smiles))
-    # print('-----', flush=True)
-    # mp_args = (m, c, xtal_lig_mols, [xtal_id], 'mcs_test.pkl', 10)
-    # rank_structures(*mp_args)
-
-    pkl.dump([compound_ids, xtal_ids, res],
-        open(f'{args.o}/mcs_sort_index.pkl', 'wb'))
-
-if __name__ == '__main__':
-    main()
